+++ packages/core-components/src/lib/atoms/index.js
@@ -1,2 +1,3 @@
 export { default as Alert } from './alert/Alert.svelte';
+export * from './skeletons';
 export { default as Dropdown } from './Dropdown.svelte';