+++ packages/plugin-connector/src/data-sources/exec-source.js
@@ -19,51 +19,50 @@ export const execSource = async (source, supportedDbs, outDir) => {
 	const runner = await db.factory(source.options, source.sourceDirectory);
 
 	console.log(`Executing ${source.name}`);
-	const results = await Promise.all(
-		source.queries.map(async (q) => {
-			const filename = q.filepath.split(path.sep).pop();
-			console.log(` >| Executing ${filename}`);
-			const before = performance.now();
-			return {
-				...q,
-				result: await runner(q.content, q.filepath).then((x) => {
-					console.log(
-						` <| Finished ${filename} (took ${(performance.now() - before).toFixed(2)}ms)`
-					);
-					return x;
-				})
-			};
-		})
-	);
-	console.log(`Finished ${source.name}`);
-
 	/** @type {Set<string>} */
 	const outputFilenames = new Set();
 
-	for (const query of results) {
-		const { result } = query;
-		if (!result) continue;
-		const parquetBuffer = await buildParquetFromResultSet(result.columnTypes, result.rows);
-		/* Split on / or \ (windows compatibility) */
-		const fileparts = query.filepath.split(/[/\\]/);
-		const outputFilename = fileparts.pop()?.split('.')[0];
+	const sourceBefore = performance.now();
+	for (const query of source.queries) {
+		const filename = query.filepath.split(path.sep).pop();
+		console.log(` >| Executing ${filename}`);
+		const before = performance.now();
+		const result = await runner(query.content, query.filepath);
+		console.log(` || Executed ${filename} (took ${(performance.now() - before).toFixed(2)}ms)`);
+		if (!result) {
+			console.log(
+				` <| Finished ${filename}. Returned no results! (took ${(
+					performance.now() - before
+				).toFixed(2)}ms)\n`
+			);
+			continue;
+		}
 
-		const outputSubdir = path.join(
-			...path
-				.join(...fileparts)
-				.split('sources')
-				.slice(1)
-		);
+		console.log(` || Writing ${filename} as parquet.`);
+		const parquetBuffer = await buildParquetFromResultSet(result.columnTypes, result.rows);
+
+		const queryDirectory = path.dirname(query.filepath);
+		const sourcesPath = path.dirname(source.sourceDirectory);
+		const outputSubdir = queryDirectory.replace(sourcesPath, '');
+
 		outputFilenames.add(
-			new URL(`file:///${path.join(outputSubdir, outputFilename + '.parquet').slice(1)}`).pathname
+			new URL(`file:///${path.join(outputSubdir, query.name + '.parquet').slice(1)}`).pathname
 		);
 		await fs.mkdir(path.join(outDir, outputSubdir), { recursive: true });
-		await fs.writeFile(path.join(outDir, outputSubdir, outputFilename + '.parquet'), parquetBuffer);
+		await fs.writeFile(path.join(outDir, outputSubdir, query.name + '.parquet'), parquetBuffer);
 		await fs.writeFile(
-			path.join(outDir, outputSubdir, outputFilename + '.schema.json'),
+			path.join(outDir, outputSubdir, query.name + '.schema.json'),
 			JSON.stringify(result.columnTypes)
 		);
+
+		console.log(
+			` <| Finished ${filename} Returned ${result?.rows.length} rows. (took ${(
+				performance.now() - before
+			).toFixed(2)}ms)\n`
+		);
 	}
+
+	console.log(`Finished ${source.name} (took ${(performance.now() - sourceBefore).toFixed(2)}ms)`);
 
 	return Array.from(outputFilenames);
 };