+++ packages/plugin-connector/src/data-sources/get-sources.js
@@ -2,8 +2,13 @@ import fs from 'fs/promises';
 import path from 'path';
 import yaml from 'yaml';
 import chalk from 'chalk';
-import { DatasourceSpecFileSchema } from './schemas/datasource-spec.schema';
+import {
+	DatasourceSpecFileSchema,
+	DatasourceCacheSchema,
+	DatasourceManifestSchema
+} from './schemas/datasource-spec.schema';
 import { cleanZodErrors } from '../lib/clean-zod-errors.js';
+import { createHash } from 'node:crypto';
 
 /**
  * Returns the path to the sources directory, if it exists in the current directory.
@@ -101,6 +106,70 @@ export const getSources = async (sourcesDir) => {
 };
 
 /**
+ *
+ * @template {import("zod").ZodType} T
+ * @param {T} zod_schema
+ * @param {string} file_path
+ * @param {import("zod").infer<T>} default_value
+ * @param {string} error_message
+ * @returns {Promise<import("zod").infer<T>>}
+ */
+async function validateFile(zod_schema, file_path, default_value, error_message) {
+	const string_default = JSON.stringify(default_value);
+
+	const file_contents = await fs.readFile(file_path, 'utf-8').catch(() => string_default);
+	const parsed = JSON.parse(file_contents);
+	const validated = zod_schema.safeParse(parsed);
+
+	if (!validated.success) {
+		console.error(chalk.bold.red(error_message));
+		await fs.writeFile(file_path, string_default);
+		return default_value;
+	}
+
+	return validated.data;
+}
+
+/**
+ *
+ * @param {string} outDir
+ * @returns {Promise<import("zod").infer<typeof DatasourceManifestSchema>>}
+ */
+export async function getCurrentManifest(outDir) {
+	const manifestPath = path.join(outDir, 'manifest.json');
+	return validateFile(
+		DatasourceManifestSchema,
+		manifestPath,
+		{ renderedFiles: {} },
+		'[!] Unable to parse manifest, ignoring'
+	);
+}
+
+const hash_location = './.evidence/template/.evidence-queries/sources/hashes.json';
+
+/**
+ * Gets the hashes of all source files, at the time of their last execution.
+ * @returns {Promise<import("zod").infer<typeof DatasourceCacheSchema>>}
+ */
+export async function getPastSourceHashes() {
+	return validateFile(
+		DatasourceCacheSchema,
+		hash_location,
+		{},
+		'[!] Unable to parse source query hashes, ignoring'
+	);
+}
+
+/**
+ * Saves the supplied source hashes
+ * @param {import("zod").infer<typeof DatasourceCacheSchema>} hashes
+ */
+export async function saveSourceHashes(hashes) {
+	await fs.mkdir(path.dirname(hash_location), { recursive: true });
+	await fs.writeFile(hash_location, JSON.stringify(hashes));
+}
+
+/**
  * Reads a YAML file containing connection parameters from the given source directory,
  * parses it, and returns a validated datasource specification.
  *
@@ -163,23 +232,75 @@ async function loadConnectionOptions(sourceDir) {
  * containing the filepath and content of each query file.
  */
 async function getQueries(sourceDir, contents) {
-	const queryFiles = contents.filter(
-		(s) => s !== 'connection.yaml' && s !== 'connection.options.yaml'
+	const queryFiles = await Promise.all(
+		contents
+			.filter((s) => s !== 'connection.yaml' && s !== 'connection.options.yaml')
+			.flatMap(
+				/**
+				 * @param {string} s
+				 * @returns {Promise<string[]>}
+				 */
+
+				async (s) => {
+					/**
+					 * @param {string} dirPath
+					 * @returns {Promise<boolean>}
+					 */
+					async function isDir(dirPath) {
+						const stats = await fs.lstat(dirPath);
+						return stats.isDirectory();
+					}
+
+					/**
+					 * @param {string} dirPath
+					 * @returns {Promise<string[]>}
+					 */
+					async function loadDirRecursive(dirPath) {
+						const content = await fs.readdir(dirPath);
+						let output = [];
+						for (const filePath of content) {
+							if (await isDir(path.join(dirPath, filePath))) {
+								output.push(...(await loadDirRecursive(path.join(dirPath, filePath))));
+							} else {
+								output.push(path.join(dirPath, filePath));
+							}
+						}
+						return output;
+					}
+
+					const fullPath = path.join(sourceDir, s);
+					if (await isDir(fullPath)) {
+						// TODO: Recurse
+						const recursed = await loadDirRecursive(fullPath);
+						return recursed.map((r) => path.relative(sourceDir, r));
+					} else {
+						return [s];
+					}
+				}
+			)
+	).then(
+		/**
+		 * @param {string[][]} r
+		 * @returns {string[]}
+		 */
+		(r) => r.flat(1)
 	);
 
 	const queries = await Promise.all(
 		queryFiles.map(async (filename) => {
 			const filepath = path.join(sourceDir, filename);
 			const { size } = await fs.stat(filepath);
-			let content;
+			let content, hash;
 			if (size > 100 * 1024 * 1024) {
 				console.warn(`${filename} is over 100MB, skipping`);
 				content = null;
+				hash = null;
 			} else {
 				content = await fs.readFile(path.join(sourceDir, filename)).then((r) => r.toString());
+				hash = createHash('md5').update(content).digest('hex');
 			}
 
-			return { filepath, content };
+			return { filepath, content, hash, name: path.basename(filepath).split('.')[0] };
 		})
 	);