+++ packages/bigquery/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @evidence-dev/bigquery
 
+## 2.0.0-usql.2
+
+### Patch Changes
+
+- 20127231: Bump all versions so version pinning works
+- Updated dependencies [20127231]
+  - @evidence-dev/db-commons@0.2.1-usql.1
+
+## 2.0.0-usql.1
+
+### Patch Changes
+
+- bdf8e08a: revamp value standardization in bigquery adapter
+- 29c149d6: added stricter types to db adapters
+- Updated dependencies [29c149d6]
+  - @evidence-dev/db-commons@0.2.1-usql.0
+
 ## 2.0.0-usql.0
 
 ### Major Changes