+++ packages/component-utilities/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @evidence-dev/component-utilities
 
+## 2.0.0-usql.7
+
+### Patch Changes
+
+- 52e114cc: move date standardization
+- ca1f90b3: Improved Logging
+
+## 2.0.0-usql.6
+
+### Patch Changes
+
+- 7c4249c0: fix falsy dates in `convertColumnToDate`
+- 20127231: Bump all versions so version pinning works
+
+## 2.0.0-usql.5
+
+### Patch Changes
+
+- 17a82581: standardize date objects in `standardizeDateString`
+
 ## 2.0.0-usql.4
 
 ### Patch Changes