+++ packages/sqlite/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @evidence-dev/sqlite
 
+## 2.0.0-usql.3
+
+### Patch Changes
+
+- 20127231: Bump all versions so version pinning works
+- Updated dependencies [20127231]
+  - @evidence-dev/db-commons@0.2.1-usql.1
+
+## 2.0.0-usql.2
+
+### Patch Changes
+
+- 29c149d6: added stricter types to db adapters
+- Updated dependencies [29c149d6]
+  - @evidence-dev/db-commons@0.2.1-usql.0
+
 ## 2.0.0-usql.1
 
 ### Patch Changes