+++ packages/tailwind/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @evidence-dev/tailwind
 
+## 1.0.0-usql.2
+
+### Patch Changes
+
+- 20127231: Bump all versions so version pinning works
+
 ## 1.0.0-usql.1
 
 ### Minor Changes