+++ packages/universal-sql/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @evidence-dev/universal-sql
 
+## 2.0.0-usql.9
+
+### Patch Changes
+
+- 52e114cc: move date standardization
+
+## 2.0.0-usql.8
+
+### Patch Changes
+
+- 20127231: Bump all versions so version pinning works
+
 ## 2.0.0-usql.7
 
 ### Patch Changes